Move db-admin and change model's folder name
