+++ app/system-modules/profile/lib/routes/password.js
@@ -13,7 +13,7 @@ var utils = Common.utils;
 
 var app = Common.app;
 
-var User = require(path.join(global.__apppath, 'model/user'));
+var User = require(path.join(global.__apppath, 'models/user'));
 
 var settings = require('../settings');
 var profileMid = require('../middleware');
